\name{data.tree.in.package-package}
\alias{data.tree.in.package-package}
\alias{data.tree.in.package}
\docType{package}
\title{\packageTitle{data.tree.in.package}}
\description{\packageDescription{data.tree.in.package}}
\section{Package Content}{\packageIndices{data.tree.in.package}}
\author{\packageAuthor{data.tree.in.package}}
\section{Maintainer}{\packageMaintainer{data.tree.in.package}}
\keyword{package}
