% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_node.R
\docType{data}
\name{test_node}
\alias{test_node}
\title{Test node}
\format{
An object of class \code{Node} (inherits from \code{R6}) of length 41.
}
\usage{
data(test_node)
}
\description{
Test node
}
\examples{
data(test_node)
print(test_node)
}
\keyword{datasets}
